Revert "Revert "5.4  最后修改合央视开始修改""
This reverts commit 8dc793abd58d17932203d611eeb04087ce6a4303.

+++ yixiupige/MODEL/JCInfoModel.cs
@@ -8,7 +8,7 @@ namespace MODEL
 {
     public class JCInfoModel
     {
-        public int jcNo { get; set; }
+        public double jcNo { get; set; }
         public int jcID { get; set; }
         public string jcCardNumber { get; set; }
         public string jcName { get; set; }
